+++ app/d_stack.py
@@ -2,11 +2,13 @@ from sanic import Sanic, response, Request, Websocket
 from sanic.log import logger
 
 from sevenball import setup_sevenball
+from streety import setup_streety
 
 
 def init():
     app = Sanic("D-Stack")
     setup_sevenball(app)
+    setup_streety(app)
     return app
